+++ src/AugmentedVR/PingpongRxBuffer.cpp
@@ -21,7 +21,7 @@ RxFrame* PingpongRxBuffer::getCurrentRxFrame(){
 void PingpongRxBuffer::put_Seq(int seq){
     NextFrameBufferPtr->RxSeq = seq;
 }
-void PingpongRxBuffer::put_TimeStamp(int ts){
+void PingpongRxBuffer::put_TimeStamp(unsigned long long ts){
     NextFrameBufferPtr->RxTimeStamp = ts;
 }
 void PingpongRxBuffer::put_TimeStamp_ZEDTS(unsigned long long ts){
